+++ src/bold.js
@@ -8,14 +8,14 @@
  */
 
 import Plugin from '@ckeditor/ckeditor5-core/src/plugin';
-import BoldEngine from './boldengine';
-import ButtonView from '@ckeditor/ckeditor5-ui/src/button/buttonview';
-import boldIcon from '../theme/icons/bold.svg';
+import BoldEditing from './bold/boldediting';
+import BoldUI from './bold/boldui';
 
 /**
- * The bold feature. It introduces the Bold button and the <kbd>Ctrl+B</kbd> keystroke.
+ * The bold feature.
  *
- * It uses the {@link module:basic-styles/boldengine~BoldEngine bold engine feature}.
+ * It loads the {@link module:basic-styles/bold/boldediting~BoldEditing bold editing feature}
+ * and {@link module:basic-styles/bold/boldui~BoldUI bold UI feature}.
  *
  * @extends module:core/plugin~Plugin
  */
@@ -24,7 +24,7 @@ export default class Bold extends Plugin {
 	 * @inheritDoc
 	 */
 	static get requires() {
-		return [ BoldEngine ];
+		return [ BoldEditing, BoldUI ];
 	}
 
 	/**
@@ -32,37 +32,5 @@ export default class Bold extends Plugin {
 	 */
 	static get pluginName() {
 		return 'Bold';
-	}
-
-	/**
-	 * @inheritDoc
-	 */
-	init() {
-		const editor = this.editor;
-		const t = editor.t;
-		const command = editor.commands.get( 'bold' );
-		const keystroke = 'CTRL+B';
-
-		// Add bold button to feature components.
-		editor.ui.componentFactory.add( 'bold', locale => {
-			const view = new ButtonView( locale );
-
-			view.set( {
-				label: t( 'Bold' ),
-				icon: boldIcon,
-				keystroke,
-				tooltip: true
-			} );
-
-			view.bind( 'isOn', 'isEnabled' ).to( command, 'value', 'isEnabled' );
-
-			// Execute command.
-			this.listenTo( view, 'execute', () => editor.execute( 'bold' ) );
-
-			return view;
-		} );
-
-		// Set the Ctrl+B keystroke.
-		editor.keystrokes.set( keystroke, 'bold' );
 	}
 }
