+++ src/italic.js
@@ -8,14 +8,14 @@
  */
 
 import Plugin from '@ckeditor/ckeditor5-core/src/plugin';
-import ItalicEngine from './italicengine';
-import ButtonView from '@ckeditor/ckeditor5-ui/src/button/buttonview';
-import italicIcon from '../theme/icons/italic.svg';
+import ItalicEditing from './italic/italicediting';
+import ItalicUI from './italic/italicui';
 
 /**
- * The italic feature. It introduces the Italic button and the <kbd>Ctrl+I</kbd> keystroke.
+ * The italic feature.
  *
- * It uses the {@link module:basic-styles/italicengine~ItalicEngine italic engine feature}.
+ * It loads the {@link module:basic-styles/italic/italicediting~ItalicEditing} and
+ * {@link module:basic-styles/italic/italicui~ItalicUI} plugins.
  *
  * @extends module:core/plugin~Plugin
  */
@@ -24,7 +24,7 @@ export default class Italic extends Plugin {
 	 * @inheritDoc
 	 */
 	static get requires() {
-		return [ ItalicEngine ];
+		return [ ItalicEditing, ItalicUI ];
 	}
 
 	/**
@@ -32,37 +32,5 @@ export default class Italic extends Plugin {
 	 */
 	static get pluginName() {
 		return 'Italic';
-	}
-
-	/**
-	 * @inheritDoc
-	 */
-	init() {
-		const editor = this.editor;
-		const t = editor.t;
-		const command = editor.commands.get( 'italic' );
-		const keystroke = 'CTRL+I';
-
-		// Add bold button to feature components.
-		editor.ui.componentFactory.add( 'italic', locale => {
-			const view = new ButtonView( locale );
-
-			view.set( {
-				label: t( 'Italic' ),
-				icon: italicIcon,
-				keystroke,
-				tooltip: true
-			} );
-
-			view.bind( 'isOn', 'isEnabled' ).to( command, 'value', 'isEnabled' );
-
-			// Execute command.
-			this.listenTo( view, 'execute', () => editor.execute( 'italic' ) );
-
-			return view;
-		} );
-
-		// Set the Ctrl+I keystroke.
-		editor.keystrokes.set( keystroke, 'italic' );
 	}
 }
