+++ src/strikethrough.js
@@ -8,14 +8,14 @@
  */
 
 import Plugin from '@ckeditor/ckeditor5-core/src/plugin';
-import StrikethroughEngine from './strikethroughengine';
-import ButtonView from '@ckeditor/ckeditor5-ui/src/button/buttonview';
-import strikethroughIcon from '../theme/icons/strikethrough.svg';
+import StrikethroughEditing from './strikethrough/strikethroughediting';
+import StrikethroughUI from './strikethrough/strikethroughui';
 
 /**
- * The strikethrough feature. It introduces the Strikethrough button and the <kbd>Ctrl+Shift+X</kbd> keystroke.
+ * The strikethrough feature.
  *
- * It uses the {@link module:basic-styles/strikethroughengine~StrikethroughEngine strikethrough engine feature}.
+ * It loads the {@link module:basic-styles/strikethrough/strikethroughediting~StrikethroughEditing} and
+ * {@link module:basic-styles/strikethrough/strikethroughui~StrikethroughUI} plugins.
  *
  * @extends module:core/plugin~Plugin
  */
@@ -24,7 +24,7 @@ export default class Strikethrough extends Plugin {
 	 * @inheritDoc
 	 */
 	static get requires() {
-		return [ StrikethroughEngine ];
+		return [ StrikethroughEditing, StrikethroughUI ];
 	}
 
 	/**
@@ -32,37 +32,5 @@ export default class Strikethrough extends Plugin {
 	 */
 	static get pluginName() {
 		return 'Strikethrough';
-	}
-
-	/**
-	 * @inheritDoc
-	 */
-	init() {
-		const editor = this.editor;
-		const t = editor.t;
-		const command = editor.commands.get( 'strikethrough' );
-		const keystroke = 'CTRL+SHIFT+X';
-
-		// Add strikethrough button to feature components.
-		editor.ui.componentFactory.add( 'strikethrough', locale => {
-			const view = new ButtonView( locale );
-
-			view.set( {
-				label: t( 'Strikethrough' ),
-				icon: strikethroughIcon,
-				keystroke,
-				tooltip: true
-			} );
-
-			view.bind( 'isOn', 'isEnabled' ).to( command, 'value', 'isEnabled' );
-
-			// Execute command.
-			this.listenTo( view, 'execute', () => editor.execute( 'strikethrough' ) );
-
-			return view;
-		} );
-
-		// Set the Ctrl+Shift+X keystroke.
-		editor.keystrokes.set( keystroke, 'strikethrough' );
 	}
 }
