+++ src/strikethrough/strikethroughediting.js
@@ -4,26 +4,26 @@
  */
 
 /**
- * @module basic-styles/strikethroughengine
+ * @module basic-styles/strikethrough/strikethroughediting
  */
 
 import Plugin from '@ckeditor/ckeditor5-core/src/plugin';
 import { downcastAttributeToElement } from '@ckeditor/ckeditor5-engine/src/conversion/downcast-converters';
 import { upcastElementToAttribute } from '@ckeditor/ckeditor5-engine/src/conversion/upcast-converters';
-import AttributeCommand from './attributecommand';
+import AttributeCommand from '../attributecommand';
 
 const STRIKETHROUGH = 'strikethrough';
 
 /**
- * The strikethrough engine feature.
+ * The strikethrough editing feature.
  *
- * It registers the `strikethrough` command and introduces the
+ * It registers the `strikethrough` command, the <kbd>Ctrl+Shift+X</kbd> keystroke and introduces the
  * `strikethroughsthrough` attribute in the model which renders to the view
  * as a `<s>` element.
  *
  * @extends module:core/plugin~Plugin
  */
-export default class StrikethroughEngine extends Plugin {
+export default class StrikethroughEditing extends Plugin {
 	/**
 	 * @inheritDoc
 	 */
@@ -46,5 +46,8 @@ export default class StrikethroughEngine extends Plugin {
 
 		// Create strikethrough command.
 		editor.commands.add( STRIKETHROUGH, new AttributeCommand( editor, STRIKETHROUGH ) );
+
+		// Set the Ctrl+Shift+X keystroke.
+		editor.keystrokes.set( 'CTRL+SHIFT+X', 'strikethrough' );
 	}
 }
